Adicioando cadastro de users do github apenas pelo admin

+++ src/Migrations/Migrations.ts
@@ -9,9 +9,18 @@ const createTables = async (drop: boolean) => {
     console.log('\n===============================');
     if (drop) {
         try {
-            await knex.schema.dropTable('users');
-            // tslint:disable-next-line: no-console
-            console.log('>>>> Tabela `users` dropada');
+            const existsTableUsers = await knex.schema.hasTable('users');
+            if (existsTableUsers) {
+                await knex.schema.dropTable('users');
+                // tslint:disable-next-line: no-console
+                console.log('>>>> Tabela `users` dropada');
+            }
+            const existsTableUsersGit = await knex.schema.hasTable('userGit');
+            if (existsTableUsersGit) {
+                await knex.schema.dropTable('userGit');
+                // tslint:disable-next-line: no-console
+                console.log('>>>> Tabela `userGit` dropada');
+            }
             return finalizar();
         } catch (error) {
             // tslint:disable-next-line: no-string-literal no-console
@@ -20,23 +29,27 @@ const createTables = async (drop: boolean) => {
         }
     }
 
-    const exists = await knex.schema.hasTable('users');
-    if (!exists) {
-        // tslint:disable-next-line: no-console
-        console.log('>>>> Tabela `users` não existe, estamos criando');
-        await knex.schema.createTable('users', (table) => {
-            table.increments('idUser').primary();
-            table.string('email', 150).notNullable().unique();
-            table.string('senha', 200).notNullable();
-            table.string('cpf', 15).notNullable().unique();
-            table.enum('tipoUsuario', ['ADMIN', 'COMUM']).notNullable();
-        });
-        // tslint:disable-next-line: no-console
-        console.log('>>>> Tabelas criada');
-        finalizar();
-    }
     // tslint:disable-next-line: no-console
-    console.log('\n>>> Todas tabelas já criadas\n');
+    console.log('>>>> Tabela `users` não existe, estamos criando');
+    await knex.schema.createTable('users', (table) => {
+        table.increments('idUser').primary();
+        table.string('email', 150).notNullable().unique();
+        table.string('senha', 200).notNullable();
+        table.string('cpf', 15).notNullable().unique();
+        table.enum('tipoUsuario', ['ADMIN', 'COMUM']).notNullable();
+    });
+    // tslint:disable-next-line: no-console
+    console.log('>>>> Tabela `userGit` não existe, estamos criando');
+    await knex.schema.createTable('userGit', (table) => {
+        table.increments('idUser').primary();
+        table.string('login', 100).notNullable().unique();
+        table.string('name', 200).notNullable();
+        table.string('bio', 1000);
+        table.string('location', 150);
+        table.string('html_url', 200).notNullable();
+    });
+    // tslint:disable-next-line: no-console
+    console.log('>>>> Tabelas criada');
     finalizar();
 };
